% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{get_alcohol}
\alias{get_alcohol}
\title{calculate drunken mass aclohol from <drinks>}
\usage{
get_alcohol(drinks)
}
\arguments{
\item{drinks}{list containing each drink and the amount of drunken glasses as numeric}
}
\value{
double of length 1 rendering the total drunken mass of alcohol in gram
}
\description{
The drunken mass of alcohol ist returned in grams and calculated as follows:
\deqn{drunken mass= \sum drinks liters alcohol}
}
