% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{get_bodywater}
\alias{get_bodywater}
\title{Calculate a Persons average bodywater}
\usage{
get_bodywater(sex = c("male", "female"), age, height, weight)
}
\arguments{
\item{sex}{character string that is either assignable to male or female}

\item{age}{number between 10 and 110 measured in years}

\item{height}{number between 100 and 230 measured in cm}

\item{weight}{number between 40 and 300 measured in kg}
}
\value{
estimated Total body water, a numeric measured in liter
}
\description{
Calculate a Persons average bodywater depending on <sex>, <age>, <height> and <weight>
}
