% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{get_permille}
\alias{get_permille}
\title{Caluculate a Persons alcohol level}
\usage{
get_permille(alcohol_drunk, bodywater, drinking_time)
}
\arguments{
\item{alcohol_drunk}{drunken mass aclohol: \code{\link{get_alcohol}}}

\item{bodywater}{persons average bodywater: \code{\link{get_bodywater}}}

\item{drinking_time}{A vector of length 2 containing two posixct, start and end of drinking}
}
\value{
blood alcohol level, a numeric measured in thousands
}
\description{
Caluculate a Persons alcohol level in thousands (permille)
}
