% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_promillo.R
\name{list_alcohol_course}
\alias{list_alcohol_course}
\title{Lists your level of drunknes during Time}
\usage{
list_alcohol_course(alcohol_drunk, bodywater, drinking_time, distance)
}
\arguments{
\item{alcohol_drunk}{drunken mass aclohol: \code{\link{get_alcohol}}}

\item{bodywater}{persons average bodywater: \code{\link{get_bodywater}}}

\item{drinking_time}{A vector of length 2 containing two posixct, start and end of drinking}

\item{distance}{determine the distance in minutes between each new measurement; default is set to 5 minutes}
}
\value{
a vector containing time and alcohollevel
}
\description{
Prepares a vector of alcohol levels in an intervall of time for plotting with \code{\link[=show_me_how_drunk]{show_me_how_drunk()}}
}
