% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_promillo.R
\name{show_me_how_drunk}
\alias{show_me_how_drunk}
\title{Vizualizes your level of drunknes during Time}
\usage{
show_me_how_drunk(
  age,
  sex,
  height,
  weight,
  drinking_time,
  drinks,
  distance = 5
)
}
\arguments{
\item{age}{number between 10 and 110 measured in years}

\item{sex}{character string that is either assignable to male or female}

\item{height}{number between 100 and 230 measured in cm}

\item{weight}{number between 40 and 300 measured in kg}

\item{drinking_time}{A vector of length 2 containing two posixct, start and end of drinking}

\item{drinks}{list containing each drink and the amount of drunken glasses as numeric}

\item{distance}{determine the distance in minutes between each new measurement; default is set to 5 minutes}
}
\value{
a ggplot-object showing the course of your alcohol intoxication
}
\description{
Plots your level of drunknes in timeintervalls
}
\seealso{
\code{\link[=tell_me_how_drunk]{tell_me_how_drunk()}} for estimation of alcohol level
}
