% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promillo.R
\name{tell_me_how_drunk}
\alias{tell_me_how_drunk}
\title{Estimates your \href{https://web.archive.org/web/20150123143123/http://promille-rechner.org/erlaeuterung-der-promille-berechnung/}{alcohol level}}
\usage{
tell_me_how_drunk(
  age,
  sex = c("male", "female"),
  height,
  weight,
  drinking_time,
  drinks
)
}
\arguments{
\item{age}{number between 10 and 110 measured in years}

\item{sex}{character string that is either assignable to male or female}

\item{height}{number between 100 and 230 measured in cm}

\item{weight}{number between 40 and 300 measured in kg}

\item{drinking_time}{A vector of length 2 containing two posixct, start and end of drinking}

\item{drinks}{list containing each drink and the amount of drunken glasses as numeric}
}
\value{
blood alcohol level, a numeric measured in thousands
}
\description{
How drunken are you? Calculate your level of drunkenness
}
